# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ArcherPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  deployed: 'deployed',
  archived: 'archived',
} as const;

export type ArcherImplementationApplicationStructure = { [key: string]: unknown };

export type ArcherImplementationModulesItem = { [key: string]: unknown };

export type ArcherImplementationFieldsItem = { [key: string]: unknown };

export type ArcherImplementationValueListsItem = { [key: string]: unknown };

export type ArcherImplementationCrossReferencesItem = { [key: string]: unknown };

export type ArcherImplementationQuestionnairesItem = { [key: string]: unknown };

export type ArcherImplementationWorkflow = { [key: string]: unknown };

export type ArcherImplementationRecordPermissionsItem = { [key: string]: unknown };

export type ArcherImplementationNotificationsItem = { [key: string]: unknown };

export type ArcherImplementationReportsItem = { [key: string]: unknown };

export type ArcherImplementationDashboardsItem = { [key: string]: unknown };

export type ArcherImplementationSampleRecordsItem = { [key: string]: unknown };

export type ArcherImplementationTestCasesItem = { [key: string]: unknown };

export interface ArcherImplementation {
  businessOverview?: string;
  useCases?: string[];
  applicationStructure?: ArcherImplementationApplicationStructure;
  modules?: ArcherImplementationModulesItem[];
  fields?: ArcherImplementationFieldsItem[];
  valueLists?: ArcherImplementationValueListsItem[];
  crossReferences?: ArcherImplementationCrossReferencesItem[];
  questionnaires?: ArcherImplementationQuestionnairesItem[];
  workflow?: ArcherImplementationWorkflow;
  recordPermissions?: ArcherImplementationRecordPermissionsItem[];
  notifications?: ArcherImplementationNotificationsItem[];
  reports?: ArcherImplementationReportsItem[];
  dashboards?: ArcherImplementationDashboardsItem[];
  sampleRecords?: ArcherImplementationSampleRecordsItem[];
  testCases?: ArcherImplementationTestCasesItem[];
}

export interface Project {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  /** @nullable */
  prompt?: string | null;
  content?: ArcherImplementation | null;
  versionCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  prompt?: string;
  content?: ArcherImplementation;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  draft: 'draft',
  deployed: 'deployed',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  content?: ArcherImplementation;
}

export interface ProjectVersion {
  id: number;
  projectId: number;
  versionNumber: number;
  content?: ArcherImplementation | null;
  createdAt: string;
}

export type ExportInputFormat = typeof ExportInputFormat[keyof typeof ExportInputFormat];


export const ExportInputFormat = {
  word: 'word',
  pdf: 'pdf',
  markdown: 'markdown',
  json: 'json',
  'archer-package': 'archer-package',
} as const;

export interface ExportInput {
  format: ExportInputFormat;
}

export interface ExportResult {
  format: string;
  filename: string;
  content: string;
  mimeType: string;
}

export interface GenerateInput {
  /** @minLength 1 */
  prompt: string;
  /** @nullable */
  projectId?: number | null;
}

export type CopilotMessageRole = typeof CopilotMessageRole[keyof typeof CopilotMessageRole];


export const CopilotMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface CopilotMessage {
  role: CopilotMessageRole;
  content: string;
}

export interface CopilotInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  projectId?: number | null;
  history?: CopilotMessage[];
}

export interface ArcherConnection {
  id: number;
  userId: string;
  name: string;
  url: string;
  username: string;
  /** @nullable */
  tenantName?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ArcherConnectionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  url: string;
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  credential: string;
  tenantName?: string;
}

export interface ArcherConnectionUpdate {
  name?: string;
  url?: string;
  username?: string;
  credential?: string;
  tenantName?: string;
  isActive?: boolean;
}

export interface ConnectionTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  details?: string | null;
}

export type DeploymentStatus = typeof DeploymentStatus[keyof typeof DeploymentStatus];


export const DeploymentStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  simulated: 'simulated',
} as const;

export type DeploymentStepStatus = typeof DeploymentStepStatus[keyof typeof DeploymentStepStatus];


export const DeploymentStepStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface DeploymentStep {
  name: string;
  status: DeploymentStepStatus;
  /** @nullable */
  message?: string | null;
}

export interface Deployment {
  id: number;
  projectId: number;
  userId: string;
  status: DeploymentStatus;
  steps: DeploymentStep[];
  /** @nullable */
  error?: string | null;
  /** @nullable */
  connectionId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface DeploymentInput {
  projectId: number;
  /** @nullable */
  connectionId?: number | null;
  simulate?: boolean;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalDeployments: number;
  savedHours: number;
  draftProjects: number;
  deployedProjects: number;
  recentActivity: ActivityItem[];
}

export interface AdminUser {
  userId: string;
  /** @nullable */
  email?: string | null;
  projectCount: number;
  deploymentCount: number;
  /** @nullable */
  lastActive?: string | null;
}

export type AdminAnalyticsProjectsPerDayItem = { [key: string]: unknown };

export type AdminAnalyticsDeploymentsPerDayItem = { [key: string]: unknown };

export interface AdminAnalytics {
  totalUsers: number;
  totalProjects: number;
  totalDeployments: number;
  projectsPerDay: AdminAnalyticsProjectsPerDayItem[];
  deploymentsPerDay: AdminAnalyticsDeploymentsPerDayItem[];
}

export interface ApiLogEntry {
  id: number;
  /** @nullable */
  userId?: string | null;
  endpoint: string;
  method: string;
  statusCode: number;
  durationMs: number;
  createdAt: string;
}

export type ListProjectsParams = {
status?: ListProjectsStatus;
search?: string;
};

export type ListProjectsStatus = typeof ListProjectsStatus[keyof typeof ListProjectsStatus];


export const ListProjectsStatus = {
  draft: 'draft',
  deployed: 'deployed',
  archived: 'archived',
} as const;

export type ListAdminLogsParams = {
limit?: number;
};