# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArcherPilot AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate:
    post:
      operationId: generateArcherImplementation
      tags: [ai]
      summary: Generate a full Archer implementation from a natural-language prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInput"
      responses:
        "200":
          description: Generated Archer implementation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArcherImplementation"
  /ai/copilot:
    post:
      operationId: sendCopilotMessage
      tags: [ai]
      summary: Send a message to the AI copilot (returns SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotInput"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  # ── Archer Connections ──────────────────────────────────────────────────────
  /archer-connections:
    get:
      operationId: listArcherConnections
      tags: [archer-connections]
      summary: List Archer connections for the authenticated user
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArcherConnection"
    post:
      operationId: createArcherConnection
      tags: [archer-connections]
      summary: Create a new Archer connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArcherConnectionInput"
      responses:
        "201":
          description: Created connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArcherConnection"
  /archer-connections/{id}:
    get:
      operationId: getArcherConnection
      tags: [archer-connections]
      summary: Get an Archer connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArcherConnection"
    put:
      operationId: updateArcherConnection
      tags: [archer-connections]
      summary: Update an Archer connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArcherConnectionUpdate"
      responses:
        "200":
          description: Updated connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArcherConnection"
    delete:
      operationId: deleteArcherConnection
      tags: [archer-connections]
      summary: Delete an Archer connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /archer-connections/{id}/test:
    post:
      operationId: testArcherConnection
      tags: [archer-connections]
      summary: Test an Archer connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"

  # ── Deployments ─────────────────────────────────────────────────────────────
...
  /admin/logs:
    get:
      operationId: listAdminLogs
      tags: [admin]
      summary: List API logs (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: API logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiLogEntry"
components:
  schemas:
...
    ArcherImplementation:
      type: object
      properties:
        businessOverview:
          type: string
        useCases:
          type: array
          items:
            type: string
        applicationStructure:
          type: object
          additionalProperties: true
        modules:
          type: array
          items:
            type: object
            additionalProperties: true
        fields:
          type: array
          items:
            type: object
            additionalProperties: true
        valueLists:
          type: array
          items:
            type: object
            additionalProperties: true
        crossReferences:
          type: array
          items:
            type: object
            additionalProperties: true
        questionnaires:
          type: array
          items:
            type: object
            additionalProperties: true
        workflow:
          type: object
          additionalProperties: true
        recordPermissions:
          type: array
          items:
            type: object
            additionalProperties: true
        notifications:
          type: array
          items:
            type: object
            additionalProperties: true
        reports:
          type: array
          items:
            type: object
            additionalProperties: true
        dashboards:
          type: array
          items:
            type: object
            additionalProperties: true
        sampleRecords:
          type: array
          items:
            type: object
            additionalProperties: true
        testCases:
          type: array
          items:
            type: object
            additionalProperties: true

    # ── AI schemas ───────────────────────────────────────────────────────────
    GenerateInput:
      type: object
      properties:
        prompt:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
      required: [prompt]
    CopilotInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
        history:
          type: array
          items:
            $ref: "#/components/schemas/CopilotMessage"
      required: [message]
    CopilotMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required: [role, content]

    # ── Archer Connection schemas ─────────────────────────────────────────────
    ArcherConnection:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        url:
          type: string
        username:
          type: string
        tenantName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, name, url, username, isActive, createdAt, updatedAt]
    ArcherConnectionInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 1
        credential:
          type: string
          minLength: 1
        tenantName:
          type: string
      required: [name, url, username, credential]
    ArcherConnectionUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        username:
          type: string
        credential:
          type: string
        tenantName:
          type: string
        isActive:
          type: boolean
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        details:
          type: ["string", "null"]
      required: [success, message]

    # ── Deployment schemas ───────────────────────────────────────────────────
...
    ApiLogEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        endpoint:
          type: string
        method:
          type: string
        statusCode:
          type: integer
        durationMs:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, endpoint, method, statusCode, durationMs, createdAt]