# lib/api-zod/src/generated/types/archerConnection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ArcherPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArcherConnection {
  id: number;
  userId: string;
  name: string;
  url: string;
  username: string;
  /** @nullable */
  tenantName?: string | null;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}