# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ArcherPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateInput {
  /** @minLength 1 */
  prompt: string;
  /** @nullable */
  projectId?: number | null;
}